+++ _tests_/react-ts-tw/src/types/user.ts
@@ -46,6 +46,4 @@ export type FullUserType = UserProfileType & {
 };
 
 // Test: discriminated union type
-export type ApiResultType<T> =
-    | { data: T, error: string, status: "error" }
-    | { data: T, error: null, status: "success" };
+export type ApiResultType = { data: T, error: string, status: "error" } | { data: T, error: null, status: "success" };